chore: move our exposure info
